move jQueryUI plugin to separate file
